add validation

+++ app/prepare.py
@@ -8,24 +8,25 @@ from config.env import env
 
 
 def kill_connection():
-    procs = list(filter(
-        lambda x: x["name"] == "python" and 0 < len(x["connections"]),
-        map(
+    processes = map(
             lambda x: x.as_dict(attrs=["name", "pid", "connections"]),
-            psutil.process_iter())))
+            psutil.process_iter())
+    if processes is not None:
+        procs = list(filter(
+            lambda x: x["name"] == "python" and 0 < len(x["connections"]), processes))
 
-    pids = {}
-    for proc in procs:
-        for connection in proc["connections"]:
-            if connection.laddr == (
-            socket.gethostbyname(env["AUTOWARE_WEB_UI_HOST"]), int(env["AUTOWARE_WEB_UI_PORT"])):
-                pids[proc["pid"]] = psutil.Process(pid=proc["pid"]);
+        pids = {}
+        for proc in procs:
+            for connection in proc["connections"]:
+                if connection.laddr == (
+                socket.gethostbyname(env["AUTOWARE_WEB_UI_HOST"]), int(env["AUTOWARE_WEB_UI_PORT"])):
+                    pids[proc["pid"]] = psutil.Process(pid=proc["pid"]);
 
-    if 0 < len(pids):
-        for pid in pids.values():
-            print(pid)
-            pid.kill()
-            pid.wait()
+        if 0 < len(pids):
+            for pid in pids.values():
+                print(pid)
+                pid.kill()
+                pid.wait()
 
 
 def kill_ros():
